add one button for registering creds
